chatbot/backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
python-frontmatter==1.0.0
sentence-transformers>=2.2.2
torch>=2.0.0
numpy>=1.21.0
qdrant-client==1.6.2
openai==0.28.1
transformers>=4.20.0